+++ src/components/OrdersBelt/OrdersBelt.tsx
@@ -6,12 +6,16 @@ import AddIcon from '@mui/icons-material/Add'
 import { useAppState } from '../../context/AppState'
 import { generatePDF } from './utils/generatePDF'
 import { generateCSV } from './utils/generateCSV'
+import { Skeleton } from '@mui/joy'
+import useRealTimeData from '../../hooks/useRealTimeData'
 
 export default function OrdersBelt(): JSX.Element {
   const {
     state: { authorizedUser, dataWithFilters },
     dispatch,
   } = useAppState()
+
+  const { rowsDataLoading } = useRealTimeData()
 
   const onClickAddNew = () =>
     dispatch({ type: 'SET_OPEN_MODAL_ADD_EDIT_ELEMENTS', payload: { modal: true, isEdit: false } })
@@ -29,9 +33,12 @@ export default function OrdersBelt(): JSX.Element {
         justifyContent: 'space-between',
       }}
     >
-      <Typography level='h2' component='h1'>
-        Orders
-      </Typography>
+      <Skeleton variant='rectangular' width={210} height={30} loading={rowsDataLoading}>
+        <Typography level='h2' component='h1'>
+          Orders
+        </Typography>
+      </Skeleton>
+
       <Box
         sx={{
           display: 'flex',
@@ -39,27 +46,35 @@ export default function OrdersBelt(): JSX.Element {
           marginLeft: 'auto',
         }}
       >
-        <Button
-          color='neutral'
-          variant='outlined'
-          startDecorator={<DownloadRoundedIcon />}
-          size='sm'
-          onClick={() => generatePDF(dataWithFilters, authorizedUser)}
-        >
-          Download PDF
-        </Button>
-        <Button
-          color='neutral'
-          variant='outlined'
-          startDecorator={<DownloadRoundedIcon />}
-          size='sm'
-          onClick={() => generateCSV(dataWithFilters)}
-        >
-          Download CSV
-        </Button>
-        <Button color='success' startDecorator={<AddIcon />} size='sm' onClick={onClickAddNew}>
-          Add new
-        </Button>
+        <Skeleton variant='rectangular' width={120} height={30} loading={rowsDataLoading}>
+          <Button
+            color='neutral'
+            variant='outlined'
+            startDecorator={<DownloadRoundedIcon />}
+            size='sm'
+            onClick={() => generatePDF(dataWithFilters, authorizedUser)}
+          >
+            Download PDF
+          </Button>
+        </Skeleton>
+
+        <Skeleton variant='rectangular' width={120} height={30} loading={rowsDataLoading}>
+          <Button
+            color='neutral'
+            variant='outlined'
+            startDecorator={<DownloadRoundedIcon />}
+            size='sm'
+            onClick={() => generateCSV(dataWithFilters)}
+          >
+            Download CSV
+          </Button>
+        </Skeleton>
+
+        <Skeleton variant='rectangular' width={120} height={30} loading={rowsDataLoading}>
+          <Button color='success' startDecorator={<AddIcon />} size='sm' onClick={onClickAddNew}>
+            Add new
+          </Button>
+        </Skeleton>
       </Box>
     </Box>
   )