+++ src/hooks/useLocalStorage.tsx
@@ -1,20 +1,30 @@
 import { useState, useEffect } from 'react'
 import { useAppState } from '../context/AppState'
+import { decryptedData, encryptedData } from '../crypto'
 
 export default function useLocalStorage(key: string, initialValue: string[]) {
   const {
     state: { authorizedUser },
   } = useAppState()
 
-  const prefixedKey = `${authorizedUser?.email}-${key}`
+  const prefixedKey = `${authorizedUser?.uid}-${key}`
 
   const [value, setValue] = useState(() => {
     const storedValue = localStorage.getItem(prefixedKey)
-    return storedValue ? JSON.parse(storedValue) : initialValue
+    if (storedValue) {
+      try {
+        const data = decryptedData(storedValue)
+        return JSON.parse(data)
+      } catch (error) {
+        return initialValue
+      }
+    }
+    return initialValue
   })
 
   useEffect(() => {
-    localStorage.setItem(prefixedKey, JSON.stringify(value))
+    const data = encryptedData(value)
+    localStorage.setItem(prefixedKey, data)
   }, [prefixedKey, value])
 
   return [value, setValue]