+++ src/ui/MenuItems/MenuItems.tsx
@@ -17,7 +17,6 @@ export default function MenuItems({ label, icon, redirect }: MenuItemsProps): JS
           border: '1px solid green',
         },
       }}
-      key={label}
     >
       <ListItemButton onClick={() => navigate(redirect)} disabled={true}>
         {icon}